+++ apps/auth/src/auth.controller.spec.ts
@@ -13,10 +13,4 @@ describe('AuthController', () => {
 
     authController = app.get<AuthController>(AuthController);
   });
-
-  describe('root', () => {
-    it('should return "Hello World!"', () => {
-      expect(authController.getHello()).toBe('Hello World!');
-    });
-  });
 });
