+++ apps/auth/src/auth.controller.ts
@@ -24,6 +24,7 @@ export class AuthController {
   @UseGuards(JwtAuthGuard)
   @MessagePattern('validate_user')
   async validateUser(@CurrentUser() user: User) {
-    return user;
+    console.log('validateUser', user);
+    return this.authService.validateUser(user);
   }
 }
