+++ apps/auth/src/auth.service.ts
@@ -3,6 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
 import { Response } from 'express';
 import { User } from './users/schemas/user.schema';
+import { UsersService } from './users/users.service';
 
 export interface TokenPayload {
   userId: string;
@@ -13,6 +14,7 @@ export class AuthService {
   constructor(
     private readonly configService: ConfigService,
     private readonly jwtService: JwtService,
+    private readonly usersService: UsersService,
   ) {}
 
   async login(user: User, response: Response) {
@@ -38,5 +40,9 @@ export class AuthService {
       httpOnly: true,
       expires: new Date(),
     });
+  }
+
+  async validateUser(user: User) {
+    return this.usersService.getUser({ _id: user._id });
   }
 }
