+++ apps/file-storage/.env
@@ -1,0 +1,4 @@
+PORT=3002
+
+RABBIT_MQ_URI=amqp://rabbitmq:5672
+RABBIT_MQ_AUTH_QUEUE=auth
