+++ apps/file-storage/src/file-storage.module.ts
@@ -1,41 +1,25 @@
-import { BadRequestException, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { FileStorageController } from './file-storage.controller';
 import { FileStorageService } from './file-storage.service';
 import { MulterModule } from '@nestjs/platform-express';
-import { diskStorage } from 'multer';
-import * as path from 'path';
-import * as fs from 'fs';
-
-const storage = diskStorage({
-  destination: function (req, file, cb) {
-    const { datasetId } = req.body;
-
-    // make sure datasetId is provided
-    if (!datasetId) {
-      cb(new BadRequestException('Missing datasetId'), null);
-      return;
-    }
-
-    // make sure datasetId is a valid folder name (no slashes, etc.)
-    if (!datasetId.match(/^[a-zA-Z0-9_-]+$/)) {
-      cb(new BadRequestException('Invalid datasetId'), null);
-      return;
-    }
-
-    const folderPath = path.join('/tmp/uploads', datasetId);
-    fs.mkdirSync(folderPath, { recursive: true });
-    cb(null, folderPath);
-  },
-  filename: function (req, file, cb) {
-    cb(null, `${Date.now()}-${file.originalname}`);
-  },
-});
+import { ConfigModule } from '@nestjs/config';
+import * as Joi from 'joi';
+import { AuthModule } from '@app/common';
+import localStorage from './storage/diskStorage';
 
 @Module({
   imports: [
-    MulterModule.register({
-      storage,
+    ConfigModule.forRoot({
+      isGlobal: true,
+      validationSchema: Joi.object({
+        PORT: Joi.number().required(),
+      }),
+      envFilePath: './apps/file-storage/.env',
     }),
+    MulterModule.register({
+      storage: localStorage,
+    }),
+    AuthModule,
   ],
   controllers: [FileStorageController],
   providers: [FileStorageService],
