+++ apps/file-storage/src/main.ts
@@ -1,9 +1,11 @@
 import { NestFactory } from '@nestjs/core';
 import { FileStorageModule } from './file-storage.module';
+import { ConfigService } from '@nestjs/config';
 
 async function bootstrap() {
   const app = await NestFactory.create(FileStorageModule);
-  await app.listen(3002);
+  const configService = app.get(ConfigService);
+  await app.listen(configService.get('PORT'));
 }
 
 bootstrap();
